#10 [refactor] : clean architecture 구조 세팅
