--- app/src/main/java/org/sopt/and/core/designsystem/component/Banner.kt
@@ -1,82 +1,0 @@
-package org.sopt.and.core.designsystem.component
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.pager.PagerState
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.unit.dp
-import org.sopt.and.R
-import org.sopt.and.ui.theme.WavveTheme
-
-@Composable
-fun Banner(pagerState: PagerState, modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier
-            .height(480.dp)
-            .padding(top = 10.dp)
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.iv_banner),
-            contentDescription = stringResource(R.string.home_banner),
-            modifier = Modifier
-                .fillMaxSize()
-                .clip(RoundedCornerShape(10.dp)),
-            contentScale = ContentScale.Crop
-        )
-
-        //gradient 레이아웃
-        Box(
-            modifier = Modifier
-                .height(480.dp)
-                .fillMaxWidth()
-                .background(
-                    brush = Brush.verticalGradient(
-                        colors = listOf(
-                            Color.Transparent,
-                            Color.Black.copy(alpha = 0.7F)
-                        )
-                    ),
-                    shape = RoundedCornerShape(10.dp)
-                )
-        )
-
-        Box(
-            modifier = Modifier
-                .align(Alignment.BottomEnd)
-                .offset(x = (-10).dp, y = (-10).dp)
-                .clip(RoundedCornerShape(20.dp))
-                .background(color = WavveTheme.colors.Gray16.copy(alpha = 0.9F))
-                .padding(horizontal = 10.dp, vertical = 5.dp),
-        ) {
-            Text(
-                buildAnnotatedString {
-                    withStyle(style = SpanStyle(color = Color.White)) {
-                        append("${pagerState.currentPage + 1}")
-                    }
-                    withStyle(style = SpanStyle(color = Color.Gray)) {
-                        append("/${pagerState.pageCount}")
-                    }
-                },
-            )
-        }
-    }
-}
