--- app/src/main/java/org/sopt/and/data/ApiFactory.kt
@@ -1,36 +1,0 @@
-package org.sopt.and.data
-
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
-import kotlinx.serialization.json.Json
-import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import org.sopt.and.BuildConfig
-import org.sopt.and.data.network.service.UserService
-import retrofit2.Retrofit
-
-object ApiFactory {
-    private const val BASE_URL: String = BuildConfig.BASE_URL
-
-    private val loggingInterceptor = HttpLoggingInterceptor().apply {
-        level = HttpLoggingInterceptor.Level.BODY
-    }
-
-    private val client = OkHttpClient.Builder()
-        .addInterceptor(loggingInterceptor)
-        .build()
-
-    val retrofit: Retrofit by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .client(client)
-            .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
-            .build()
-    }
-
-    inline fun <reified T> create(): T = retrofit.create(T::class.java)
-}
-
-object ServicePool {
-    val userService = ApiFactory.create<UserService>()
-}
