+++ app/src/main/java/org/sopt/and/data/mapper/todata/UserLoginModelMapper.kt
@@ -1,0 +1,7 @@
+package org.sopt.and.data.mapper.todata
+
+import org.sopt.and.domain.model.request.UserLoginModel
+
+fun UserLoginModel.toData() : UserLoginRequestDto = UserLoginRequestDto(
+
+)
