+++ app/src/main/java/org/sopt/and/data/mapper/todata/UserSignUpModelMapper.kt
@@ -1,0 +1,7 @@
+package org.sopt.and.data.mapper.todata
+
+import org.sopt.and.domain.model.request.UserSignUpModel
+
+fun UserSignUpModel.toData() : UserSignUpRequestDto = UserSignUpRequestDto(
+
+)
