+++ app/src/main/java/org/sopt/and/data/mapper/todomain/UserHobbyResponseDtoMapper.kt
@@ -1,0 +1,7 @@
+package org.sopt.and.data.mapper.todomain
+
+import org.sopt.and.domain.model.response.Hobby
+
+fun UserHobbyResponseDto.toDomain(): Hobby = Hobby(
+    hobby = this.hobby
+)
