+++ app/src/main/java/org/sopt/and/data/remote/datasource/UserDataSource.kt
@@ -1,0 +1,7 @@
+package org.sopt.and.data.remote.datasource
+
+interface UserDataSource {
+    suspend fun postUserSignUp() : ApiResponse<UserSignUpResponseDto>
+    suspend fun postUserLogin() : ApiResponse<UserTokenResponseDto>
+    suspend fun getUserHobby() : ApiResponse<UserHobbyResponseDto>
+}
