+++ app/src/main/java/org/sopt/and/data/remote/datasourceimpl/UserDataSourceImpl.kt
@@ -1,0 +1,11 @@
+package org.sopt.and.data.remote.datasourceimpl
+
+import org.sopt.and.data.remote.datasource.UserDataSource
+
+class UserDataSourceImpl @Inject constructor(
+  //  private val
+) : UserDataSource {
+    override suspend fun getUserHobby(): ApiResponse<UserHobbyResponseDto> {
+        userService.
+    }
+}
