+++ app/src/main/java/org/sopt/and/domain/model/request/UserLoginModel.kt
@@ -1,0 +1,6 @@
+package org.sopt.and.domain.model.request
+
+data class UserLoginModel(
+    val username: String,
+    val password: String
+)
