+++ app/src/main/java/org/sopt/and/domain/model/request/UserSignUpModel.kt
@@ -1,0 +1,7 @@
+package org.sopt.and.domain.model.request
+
+data class UserSignUpModel(
+    val username: String,
+    val password: String,
+    val hobby: String
+)
