+++ app/src/main/java/org/sopt/and/domain/model/response/Hobby.kt
@@ -1,0 +1,5 @@
+package org.sopt.and.domain.model.response
+
+data class Hobby(
+    val hobby: String
+)
