+++ app/src/main/java/org/sopt/and/domain/model/response/Token.kt
@@ -1,0 +1,5 @@
+package org.sopt.and.domain.model.response
+
+data class Token(
+    val token: String
+)
