+++ app/src/main/java/org/sopt/and/domain/model/response/UserNumber.kt
@@ -1,0 +1,5 @@
+package org.sopt.and.domain.model.response
+
+data class UserNumber(
+    val no: Int
+)
