+++ app/src/main/java/org/sopt/and/domain/repository/UserHobbyRepository.kt
@@ -1,0 +1,4 @@
+package org.sopt.and.domain.repository
+
+interface UserHobbyRepository {
+}
