--- app/src/main/java/org/sopt/and/feature/login/LoginScreen.kt
@@ -1,245 +1,0 @@
-package org.sopt.and.feature.login
-
-import android.content.Context
-import androidx.activity.compose.LocalOnBackPressedDispatcherOwner
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.rounded.KeyboardArrowLeft
-import androidx.compose.material3.CenterAlignedTopAppBar
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.VerticalDivider
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.livedata.observeAsState
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusDirection
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavController
-import org.sopt.and.R
-import org.sopt.and.core.designsystem.component.AuthTextField
-import org.sopt.and.core.designsystem.component.SocialLoginButtonGroup
-import org.sopt.and.core.designsystem.component.WavveLoginButton
-import org.sopt.and.data.remote.model.request.UserLoginRequestDto
-import org.sopt.and.feature.main.Routes
-import org.sopt.and.ui.theme.WavveTheme
-import org.sopt.and.utils.noRippleClickable
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun LoginScreen(
-    navController: NavController,
-    onLoginSuccess: (String, String) -> Unit,
-    viewModel: LoginViewModel = viewModel()
-) {
-    val context = LocalContext.current
-    val focusManager = LocalFocusManager.current
-    val dispatcher = LocalOnBackPressedDispatcherOwner.current?.onBackPressedDispatcher ?: return
-    val sharedPreferences = context.getSharedPreferences("token", Context.MODE_PRIVATE)
-    val editor = sharedPreferences.edit()
-
-
-    val token by viewModel.token.observeAsState("")
-    val userName by viewModel.userName.observeAsState("")
-    val password by viewModel.password.observeAsState("")
-    val showPassword = remember { mutableStateOf(false) }
-
-    Scaffold(
-        modifier = Modifier.fillMaxSize(),
-        topBar = {
-            CenterAlignedTopAppBar(
-                title = {
-                    Image(
-                        painter = painterResource(id = R.drawable.iv_wavve_logo),
-                        contentDescription = stringResource(R.string.wavve_logo),
-                        modifier = Modifier
-                            .size(90.dp)
-                    )
-                },
-                navigationIcon = {
-                    IconButton(
-                        onClick = {
-                            dispatcher.onBackPressed()
-                        }
-                    ) {
-                        Icon(
-                            Icons.AutoMirrored.Rounded.KeyboardArrowLeft,
-                            contentDescription = stringResource(R.string.back),
-                            tint = Color.White,
-                            modifier = Modifier
-                                .size(40.dp)
-                        )
-                    }
-                },
-                colors = TopAppBarDefaults.mediumTopAppBarColors(containerColor = WavveTheme.colors.BackgroundGray)
-            )
-        },
-    ) { innerPadding ->
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(color = WavveTheme.colors.BackgroundGray)
-                .padding(innerPadding)
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .padding(horizontal = 20.dp)
-            ) {
-                Spacer(modifier = Modifier.height(20.dp))
-
-                AuthTextField(
-                    modifier = Modifier
-                        .fillMaxWidth(),
-                    value = userName,
-                    onValueChange = {
-                        viewModel.setUserName(it)
-                    },
-                    placeholder = stringResource(R.string.placeholder_user_name),
-                    keyboardOptions = KeyboardOptions.Default.copy(
-                        keyboardType = KeyboardType.Email,
-                        imeAction = ImeAction.Next
-                    ),
-                    keyboardActions = KeyboardActions(
-                        onNext = { focusManager.moveFocus(FocusDirection.Next) }
-                    ),
-                )
-
-                Spacer(modifier = Modifier.height(5.dp))
-
-                AuthTextField(
-                    modifier = Modifier
-                        .fillMaxWidth(),
-                    value = password,
-                    onValueChange = {
-                        viewModel.setPassword(it)
-                    },
-                    placeholder = stringResource(R.string.placeholder_password),
-                    suffix = {
-                        Text(
-                            if (showPassword.value) stringResource(R.string.hide) else stringResource(
-                                R.string.show
-                            ),
-                            color = Color.White,
-                            modifier = Modifier.noRippleClickable {
-                                showPassword.value = !showPassword.value
-                            },
-                        )
-                    },
-                    keyboardOptions = KeyboardOptions.Default.copy(
-                        keyboardType = KeyboardType.Password,
-                        imeAction = ImeAction.Done
-                    ),
-                    keyboardActions = KeyboardActions(
-                        onDone = {
-                            focusManager.clearFocus()
-                        }
-                    ),
-                    visualTransformation = if (showPassword.value) VisualTransformation.None else PasswordVisualTransformation(),
-                )
-
-                Spacer(modifier = Modifier.height(30.dp))
-
-                //기본 로그인 버튼
-                WavveLoginButton(
-                    onClick = {
-                        viewModel.postUserLogin(
-                            context = context,
-                            body = UserLoginRequestDto(
-                                username = userName,
-                                password = password
-                            ),
-                        )
-
-                        editor.putString(context.getString(R.string.login_token), token)
-                        editor.apply()
-                        onLoginSuccess(userName, password)
-                        //키보드 내리기
-                        focusManager.clearFocus()
-                    }
-                )
-
-                Spacer(modifier = Modifier.height(20.dp))
-
-                Box(
-                    modifier = Modifier.fillMaxWidth(),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Row(
-                        horizontalArrangement = Arrangement.spacedBy(10.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                        modifier = Modifier.height(IntrinsicSize.Min)
-                    ) {
-                        Text(
-                            stringResource(R.string.find_id),
-                            fontSize = 12.sp,
-                            color = Color.Gray,
-                        )
-                        VerticalDivider(
-                            thickness = 1.dp,
-                            color = Color.DarkGray
-                        )
-                        Text(
-                            stringResource(R.string.reset_password),
-                            fontSize = 12.sp,
-                            color = Color.Gray,
-                        )
-                        VerticalDivider(
-                            thickness = 1.dp,
-                            color = Color.DarkGray
-                        )
-                        Text(
-                            stringResource(R.string.sign_up),
-                            fontSize = 12.sp,
-                            color = Color.Gray,
-                            modifier = Modifier.clickable {
-                                navController.navigate(Routes.SignUp.screen) {
-                                    popUpTo(Routes.SignUp.screen) { inclusive = true }
-                                }
-                            }
-                        )
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(40.dp))
-
-                //소셜 로그인 버튼
-                SocialLoginButtonGroup(stringResource(R.string.social_description))
-            }
-        }
-
-    }
-}
