--- app/src/main/java/org/sopt/and/feature/main/MainActivity.kt
@@ -1,43 +1,0 @@
-package org.sopt.and.feature.main
-
-import android.annotation.SuppressLint
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Scaffold
-import androidx.compose.ui.Modifier
-import androidx.navigation.compose.rememberNavController
-import org.sopt.and.ui.theme.ANDANDROIDTheme
-import org.sopt.and.utils.currentRoute
-
-class MainActivity : ComponentActivity() {
-
-    @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            ANDANDROIDTheme {
-                val navController = rememberNavController()
-                val isBottomAppBarVisible = navController.currentRoute() in listOf(
-                    Routes.Home.screen,
-                    Routes.Search.screen,
-                    Routes.My.screen,
-                )
-
-                Scaffold(
-                    modifier = Modifier.fillMaxSize(),
-                    bottomBar = {
-                        if (isBottomAppBarVisible) {
-                            BottomNavigationBar(navController)
-                        }
-                    }
-                ) {
-                    NavGraph(navController)
-                }
-            }
-        }
-    }
-}
