--- app/src/main/java/org/sopt/and/feature/search/SearchViewModel.kt
@@ -1,6 +1,0 @@
-package org.sopt.and.feature.search
-
-import androidx.lifecycle.ViewModel
-
-class SearchViewModel : ViewModel() {
-}
